+++ scurrypy/core/model.py
@@ -1,19 +1,15 @@
 from dataclasses import dataclass, fields, is_dataclass
 from typing import get_args, get_origin, Union
 
-def unwrap_optional(t):
-    """Extract the type from Optional[t]."""
-    if get_origin(t) is Union:
-        args = tuple(a for a in get_args(t) if a is not type(None))
-        return args[0] if len(args) == 1 else Union[args]
-    return t
+"""Extract the type from Optional[t]."""
+unwrap_optional = lambda t: get_args(t)[0] if get_origin(t) is Union else t
 
 @dataclass
 class DataModel:    
     """DataModel is a base class for Discord JSONs that provides 
         hydration from raw dicts, and optional field defaults.
     """
-    
+
     @classmethod
     def from_dict(cls, data: dict):
         """Hydrates the given data into the dataclass.
@@ -24,46 +20,44 @@ class DataModel:
         Returns:
             (cls): hydrated dataclass
         """
-        
         if not data:
             return None
-
-        kwargs = {}
-
-        for f in fields(cls):
-            key = f.name
-            value = data.get(key)
-            type_ = unwrap_optional(f.type)
-
+        
+        def convert(t, v):
+            t = unwrap_optional(t)
+            o = get_origin(t)
+            
             # missing field
-            if value is None:
-                kwargs[key] = None
-                continue
-
-            # nested dataclass
-            if is_dataclass(type_):
-                kwargs[key] = type_.from_dict(value)
-                continue
-
-            # list field
-            if get_origin(type_) is list:
-                element_type = get_args(type_)[0]
-                if is_dataclass(element_type):
-                    kwargs[key] = [
-                        element_type.from_dict(x) for x in value
-                    ]
-                else:
-                    kwargs[key] = [element_type(x) for x in value]
-                continue
-
-            # primitive field
-            try:
-                kwargs[key] = type_(value)
-            except Exception:
-                kwargs[key] = value
+            if v is None:
+                return None
+            
+            if t is bool:
+                return v == 'true'
+            
+            if is_dataclass(t):
+                return t.from_dict(v)
+            
+            if o is dict:
+                vt = get_args(t)[1]
+                return {
+                    int(k): convert(vt, x) 
+                    for k, x in v.items()
+                }
+            
+            if o is list:
+                lt = get_args(t)[0]
+                return [convert(lt, x) for x in v]
+            
+            # primitive / fallback
+            return t(v)
+        
+        kwargs = {
+            f.name: convert(f.type, data.get(f.name)) 
+            for f in fields(cls)
+        }
 
         return cls(**kwargs)
-    
+        
     def to_dict(self):
         """Recursively turns the dataclass into a dictionary and drops empty fields.